добавить statisticsService к ServicesAssembly
